ai-service/requirements.txt
# AI Service Dependencies
flask==3.0.0
flask-cors==4.0.0
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
gunicorn>=21.0.0
